refactor: modularize codebase by splitting styles and scripts into component-based files and reorganizing project structure
