+++ _archive/stitch/DESIGN.md
