+++ css/base.css
@@ -1,0 +1,160 @@
+/* LearnOps Tamagotchi Core Stylesheet */
+
+/* ┌──────────────────────────────────────────────────────────┐
+   │  CSS TABLE OF CONTENTS                                   │
+   │  1. @property registrations                              │
+   │  2. Primitive tokens + design system tokens              │
+   │  3. Base elements (typography, buttons, controls)        │
+   │  4. Grid container + tile styles                         │
+   │  5. Crystal instance state selectors                     │
+   │  6. Night mode (.night on body)                          │
+   │  7. @keyframes                                           │
+   │  8. Transition animation classes                         │
+   │  9. Sidebar + backdrop                                   │
+   │  10. Concept list items                                  │
+   │  11. Content input overlay                               │
+   │  12. Drawer creation form                                │
+   │  13. App layout (sidebar + hero, responsive)             │
+   │  14. Settings panel                                      │
+   └──────────────────────────────────────────────────────────┘ */
+
+
+/* ── 3. Base elements (typography, buttons, controls) ─────── */
+*, *::before, *::after { box-sizing:border-box; margin:0; padding:0; }
+
+body {
+  font-family: var(--font-body);
+  background-color: var(--surface);
+  color: var(--text);
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  min-height: 100vh;
+  transition: background-color 0.6s ease;
+}
+
+/* Mobile fallback card (overridden at desktop in section 13) */
+#card {
+  background-color: transparent;
+  text-align: center;
+  max-width: 440px;
+  width: 100%;
+  position: relative;
+  transition: opacity 0.3s ease, transform 0.3s ease;
+}
+body[data-drawer-open="true"] #card {
+  opacity: 0.45;
+  transform: scale(0.97);
+  pointer-events: none;
+}
+
+h2 {
+  font-family: var(--font-display);
+  font-size: 10.5px;
+  font-weight: 700;
+  letter-spacing: 1px;
+  text-transform: uppercase;
+  color: var(--text-sub);
+  margin: 0 0 4px;
+  transition: color 0.6s ease;
+}
+p.desc {
+  font-family: var(--font-display);
+  color: var(--text);
+  font-size: 22px;
+  font-weight: 700;
+  letter-spacing: -0.5px;
+  line-height: 1.3;
+  min-height: 36px;
+  margin: 8px 0 28px;
+  transition: color 0.6s ease;
+}
+.concept-label {
+  font-family: var(--font-display);
+  font-size: 13px;
+  font-weight: 600;
+  letter-spacing: 0.3px;
+  color: var(--primary);
+  margin-bottom: 4px;
+  min-height: 14px;
+  transition: color 0.6s ease;
+}
+
+.controls { display:flex; flex-direction:column; gap:10px; }
+.row      { display:flex; gap:10px; }
+
+button {
+  flex:1; padding:10px 16px; border:none; border-radius:var(--radius-btn);
+  font-family:var(--font-body); font-size:13px; font-weight:600; cursor:pointer;
+  background-color:var(--primary); color:white; letter-spacing:0.2px;
+  transition: all 0.2s cubic-bezier(0.2, 0.8, 0.2, 1);
+  box-shadow: 0 4px 10px rgba(124, 111, 205, 0.2);
+}
+button:hover:not(:disabled) { opacity: 0.95; transform: translateY(-1px); box-shadow: 0 6px 14px rgba(124, 111, 205, 0.3); }
+button:active:not(:disabled) { transform: translateY(1px); box-shadow: 0 2px 4px rgba(124, 111, 205, 0.2); }
+button:disabled { background-color:var(--locked); cursor:not-allowed; transform:none; opacity:1; }
+.btn-fail { background-color:var(--danger); }
+.btn-pass { background-color:var(--success); }
+
+#timer-display {
+  font-size:26px; font-family:'SF Mono','Fira Code',monospace;
+  font-weight:700; letter-spacing:2px; color:var(--primary);
+  margin:0 0 20px; display:none; transition:color 0.6s ease;
+}
+
+.drawer-toggle {
+  background:none; border:none; cursor:pointer;
+  padding:6px; border-radius:8px; color:var(--text-sub);
+  font-size:18px; line-height:1; flex:none; width:auto;
+  transition:color 0.2s, background-color 0.2s;
+}
+.drawer-toggle:hover { color:var(--primary); background-color:var(--border); }
+
+.dev-skip {
+  background:none; color:transparent; font-size:9px;
+  padding:4px 6px; box-shadow:none; width:auto; flex:none; letter-spacing:0;
+}
+.dev-skip:hover { color:var(--text-sub) !important; }
+
+/* ── 7. @keyframes ───────────────────────────────────────── */
+@keyframes hibernatePulse {
+  0%,100% { opacity:0.15; }
+  50%      { opacity:0.45; }
+}
+@keyframes gridEmerge {
+  from { opacity:0.2; filter:blur(1px); }
+  to   { opacity:1;   filter:none; }
+}
+@keyframes gridCrack {
+  0%   { filter:brightness(1); }
+  20%  { filter:brightness(1.6) saturate(0.3); }
+  50%  { filter:brightness(0.8); }
+  100% { filter:brightness(1); }
+}
+@keyframes gridCocoon {
+  0%   { filter:brightness(1); }
+  40%  { filter:brightness(1.3) blur(0.5px); }
+  100% { filter:brightness(1) blur(0); }
+}
+@keyframes gridActualize {
+  0%   { filter:brightness(0.8); }
+  40%  { filter:brightness(1.5) saturate(1.2); }
+  100% { filter:brightness(1); }
+}
+@keyframes gridRepair {
+  0%   { filter:brightness(0.8) saturate(0.5); }
+  50%  { filter:brightness(1.3) saturate(1.2); }
+  100% { filter:brightness(1) saturate(1); }
+}
+@keyframes tileSelect {
+  from { opacity:0; transform:scale(0.95); }
+  to   { opacity:1; transform:scale(1); }
+}
+
+/* ── 8. Transition animation classes ─────────────────────── */
+.crystal-anim.anim-emerge    { animation:gridEmerge    500ms ease forwards; }
+.crystal-anim.anim-crack     { animation:gridCrack     650ms ease forwards; }
+.crystal-anim.anim-cocoon    { animation:gridCocoon    700ms ease forwards; }
+.crystal-anim.anim-actualize { animation:gridActualize 750ms ease forwards; }
+.crystal-anim.anim-repair    { animation:gridRepair    550ms ease forwards; }
+
