+++ css/crystal.css
@@ -1,0 +1,124 @@
+/* ── 4. Grid container + tile styles ─────────────────────── */
+#grid-container {
+  width:280px;
+  margin: 10px auto 16px;
+}
+#grid-svg {
+  display:block;
+  overflow:visible;
+}
+.tile-group {
+  pointer-events:all;
+}
+
+/* Tile polygons */
+.tile-left  { fill:var(--tile-left);  transition:fill 0.6s ease; }
+.tile-right { fill:var(--tile-right); transition:fill 0.6s ease; }
+.tile-top   { fill:var(--tile-top);   transition:fill 0.6s ease; }
+
+/* Click-catch polygon on each tile */
+.tile-hit {
+  fill:transparent;
+  cursor:pointer;
+  pointer-events:visiblePainted;
+}
+
+/* Selection highlight */
+.tile-highlight {
+  fill:none;
+  stroke:var(--primary);
+  stroke-width:2.5;
+  opacity:0;
+  transition:opacity 0.25s ease, stroke 0.6s ease;
+  pointer-events:none;
+}
+.tile-group.selected .tile-highlight { opacity:1; }
+
+/* Empty slot */
+.tile-group.empty .tile-left,
+.tile-group.empty .tile-right { opacity:0.35; }
+.tile-group.empty .tile-top   { fill:none; }
+.tile-group.empty .tile-top-dash {
+  fill:none;
+  stroke:var(--locked);
+  stroke-width:1.5;
+  stroke-dasharray:4 3;
+  opacity:0.4;
+}
+.tile-group.empty .tile-hit { cursor:default; }
+
+/* Crystal instance polygons — CSS transitions for color changes */
+.crystal-instance polygon {
+  transition: fill var(--morph-ms) ease, opacity var(--morph-ms) ease;
+}
+
+/* Animation wrapper — receives keyframe animation classes */
+.crystal-anim { transition: opacity 0.3s ease; }
+
+/* ── 5. Crystal state color selectors ────────────────────── */
+.crystal-instance[data-state="instantiated"] .cp-top         { fill:hsl(270,20%,68%); }
+.crystal-instance[data-state="instantiated"] .cp-upper-left  { fill:hsl(270,18%,55%); }
+.crystal-instance[data-state="instantiated"] .cp-upper-right { fill:hsl(270,16%,46%); }
+.crystal-instance[data-state="instantiated"] .cp-lower-left  { fill:hsl(270,16%,46%); }
+.crystal-instance[data-state="instantiated"] .cp-lower-right { fill:hsl(270,14%,38%); }
+.crystal-instance[data-state="instantiated"] .cp-bottom-tip  { fill:hsl(270,14%,38%); }
+.crystal-instance[data-state="instantiated"] .cp-specular    { fill:hsl(270,30%,85%); opacity:0.35; }
+.crystal-instance[data-state="instantiated"] .cp-glow        { fill:hsl(270,20%,60%); opacity:0.10; }
+
+.crystal-instance[data-state="growing"] .cp-top         { fill:hsl(280,45%,72%); }
+.crystal-instance[data-state="growing"] .cp-upper-left  { fill:hsl(275,42%,60%); }
+.crystal-instance[data-state="growing"] .cp-upper-right { fill:hsl(270,40%,50%); }
+.crystal-instance[data-state="growing"] .cp-lower-left  { fill:hsl(270,40%,46%); }
+.crystal-instance[data-state="growing"] .cp-lower-right { fill:hsl(265,38%,38%); }
+.crystal-instance[data-state="growing"] .cp-bottom-tip  { fill:hsl(265,38%,35%); }
+.crystal-instance[data-state="growing"] .cp-specular    { fill:hsl(285,60%,90%); opacity:0.70; }
+.crystal-instance[data-state="growing"] .cp-glow        { fill:hsl(275,45%,65%); opacity:0.20; }
+
+.crystal-instance[data-state="fractured"] .cp-top         { fill:hsl(275,28%,62%); }
+.crystal-instance[data-state="fractured"] .cp-upper-left  { fill:hsl(275,25%,50%); }
+.crystal-instance[data-state="fractured"] .cp-upper-right { fill:hsl(255,32%,44%); }
+.crystal-instance[data-state="fractured"] .cp-lower-left  { fill:hsl(275,22%,42%); }
+.crystal-instance[data-state="fractured"] .cp-lower-right { fill:hsl(255,28%,35%); }
+.crystal-instance[data-state="fractured"] .cp-bottom-tip  { fill:hsl(260,28%,32%); }
+.crystal-instance[data-state="fractured"] .cp-specular    { fill:hsl(275,20%,78%); opacity:0.28; }
+.crystal-instance[data-state="fractured"] .cp-glow        { fill:hsl(290,30%,50%); opacity:0.07; }
+
+.crystal-instance[data-state="hibernating"] .cp-top         { fill:hsl(260,50%,55%); }
+.crystal-instance[data-state="hibernating"] .cp-upper-left  { fill:hsl(258,48%,40%); }
+.crystal-instance[data-state="hibernating"] .cp-upper-right { fill:hsl(255,45%,32%); }
+.crystal-instance[data-state="hibernating"] .cp-lower-left  { fill:hsl(258,45%,28%); }
+.crystal-instance[data-state="hibernating"] .cp-lower-right { fill:hsl(255,42%,22%); }
+.crystal-instance[data-state="hibernating"] .cp-bottom-tip  { fill:hsl(255,42%,20%); }
+.crystal-instance[data-state="hibernating"] .cp-specular    { fill:hsl(270,60%,80%); opacity:0.50; }
+.crystal-instance[data-state="hibernating"] .cp-glow {
+  fill:hsl(260,55%,60%);
+  animation:hibernatePulse 3s ease-in-out infinite;
+}
+
+.crystal-instance[data-state="actualized"] .cp-top         { fill:hsl(282,48%,78%); }
+.crystal-instance[data-state="actualized"] .cp-upper-left  { fill:hsl(278,45%,65%); }
+.crystal-instance[data-state="actualized"] .cp-upper-right { fill:hsl(272,42%,54%); }
+.crystal-instance[data-state="actualized"] .cp-lower-left  { fill:hsl(272,42%,50%); }
+.crystal-instance[data-state="actualized"] .cp-lower-right { fill:hsl(268,40%,42%); }
+.crystal-instance[data-state="actualized"] .cp-bottom-tip  { fill:hsl(268,40%,38%); }
+.crystal-instance[data-state="actualized"] .cp-specular    { fill:white; opacity:0.88; }
+.crystal-instance[data-state="actualized"] .cp-glow        { fill:hsl(278,50%,70%); opacity:0.30; }
+
+/* ── 6. Night mode ────────────────────────────────────────── */
+body.night {
+  --bg:           #12111e;
+  --card-bg:      #1c1a2e;
+  --text:         #ccc8e8;
+  --text-sub:     #7b78a0;
+  --primary:      #9d8fe8;
+  --locked:       #3a3756;
+  --drawer-bg:    #161428;
+  --border:       rgba(160,140,240,0.10);
+  --tile-top:     hsl(260,20%,18%);
+  --tile-left:    hsl(260,15%,13%);
+  --tile-right:   hsl(260,15%,10%);
+  --surface:      #12111e;
+  --surface-low:  #161428;
+  --surface-card: #1c1a2e;
+}
+
