+++ js/geo.js
@@ -1,0 +1,65 @@
+// js/geo.js
+
+// Index order: [0]cp-top [1]cp-upper-left [2]cp-upper-right
+//              [3]cp-lower-left [4]cp-lower-right
+//              [5]cp-bottom-tip [6]cp-specular [7]cp-glow
+export const GEO = {
+  instantiated: [
+    [100,90,  85,108, 100,130, 115,108],
+    [100,90,  85,108,  82,130, 100,130],
+    [100,90, 100,130, 118,130, 115,108],
+    [100,130,  82,130,  88,152, 100,168],
+    [100,130, 100,168, 112,152, 118,130],
+    [88,152, 100,168, 112,152],
+    [106,96, 105,112, 109,103],
+    [100,90, 115,108, 118,130, 112,152, 100,168, 88,152, 82,130, 85,108],
+  ],
+  growing: [
+    [100,73,  79,91, 100,119, 121,91],
+    [100,73,  79,91,  69,119, 100,119],
+    [100,73, 100,119, 131,119, 121,91],
+    [100,119,  69,119,  83,145, 100,167],
+    [100,119, 100,167, 117,145, 131,119],
+    [83,145, 100,167, 117,145],
+    [104,77, 114,94, 112,85],
+    [100,73, 121,91, 131,119, 117,145, 100,167, 83,145, 69,119, 79,91],
+  ],
+  fractured: [
+    [101,73,  77,93, 101,119, 123,89],
+    [101,73,  77,93,  66,120, 101,119],
+    [101,73, 101,119, 134,117, 123,89],
+    [101,119,  66,120,  82,143, 100,168],
+    [101,119, 100,168, 116,147, 134,117],
+    [82,143, 100,168, 116,147],
+    [105,76, 117,91, 114,83],
+    [101,73, 123,89, 134,117, 116,147, 100,168, 82,143, 66,120, 77,93],
+  ],
+  hibernating: [
+    [100,38,  64,70, 100,118, 136,70],
+    [100,38,  64,70,  48,118, 100,118],
+    [100,38, 100,118, 152,118, 136,70],
+    [100,118,  48,118,  72,165, 100,205],
+    [100,118, 100,205, 128,165, 152,118],
+    [72,165, 100,205, 128,165],
+    [106,44, 126,74, 120,58],
+    [100,38, 136,70, 152,118, 128,165, 100,205, 72,165, 48,118, 64,70],
+  ],
+  actualized: [
+    [100,22,  52,60, 100,118, 148,60],
+    [100,22,  52,60,  32,118, 100,118],
+    [100,22, 100,118, 168,118, 148,60],
+    [100,118,  32,118,  62,172, 100,215],
+    [100,118, 100,215, 138,172, 168,118],
+    [62,172, 100,215, 138,172],
+    [106,28, 138,64, 126,45],
+    [100,22, 148,60, 168,118, 138,172, 100,215, 62,172, 32,118, 52,60],
+  ],
+};
+
+export function easeInOutCubic(t) { return t < 0.5 ? 4*t*t*t : 1-Math.pow(-2*t+2,3)/2; }
+export function interpCoords(from, to, t) { return from.map((v,i)=>v+(to[i]-v)*t); }
+export function coordsToPoints(arr) {
+  const pts = [];
+  for (let i=0; i<arr.length; i+=2) pts.push(arr[i]+','+arr[i+1]);
+  return pts.join(' ');
+}
